ADD A NEW HEADER FILE WHICH DEFINES NMEA 2000 STANDARD ENUM TYPES.
Add a new header file (src/NMEA2000StdTypes.h) which defines data
types using the language of the NMEA2000 specification.  The NMEA
2000 standard defines data types with identifiers of the form
"DDnnn" and delivers a consistent and audited collection of types.
Although the standard documentation is not in the public domain,
there are many associated documents which expose the standard
types.

Additionally, include this file into the existing src/N2kTypes.h,
extending compatability with the standard. Aliasing of the
standard types into the more discursive naming style already
adopted in N2kTypes.h is further possible using the long names
defined by the NMEA 2000 standard for each type.

This proposal includes a "starter" definition for
src/NMEA2000StdTypes.h and an updated src/N2kTypes.h which together
illustrate this proposal.

These new files include a collection of enum definitions relating
to PGNs

- 128776 Windlass Control Status;
- 128777 Windlass Operating Status;
- 128778 Windlass Monitoring Status.

Mostly these types are peculiar to the windlass domain, but some
are more generic and have not been previously defined as such in
the existing version of src/N2kTypes.h.
